requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Auth & sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Twilio (OTP SMS)
twilio==9.3.6

# Claude API
anthropic==0.37.1

# Scraping
beautifulsoup4==4.12.3
playwright==1.47.0
httpx==0.27.2

# Planificateur de tâches
apscheduler==3.10.4

# Config
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
